Refactor personality imports and package structure
Consolidated the import paths for personality-related enums by renaming directories and updating package paths in multiple files. This refactoring aims to improve the organization and clarity of the project's structure.
