+++ MekHQ/src/mekhq/gui/dialog/CustomizePersonDialog.java
@@ -38,8 +38,8 @@ import mekhq.campaign.Campaign;
 import mekhq.campaign.personnel.*;
 import mekhq.campaign.personnel.enums.Phenotype;
 import mekhq.campaign.personnel.enums.education.EducationLevel;
-import mekhq.campaign.personnel.enums.randomEvents.personalities.*;
 import mekhq.campaign.personnel.randomEvents.PersonalityController;
+import mekhq.campaign.personnel.randomEvents.enums.personalities.*;
 import mekhq.campaign.unit.Unit;
 import mekhq.campaign.universe.Faction;
 import mekhq.campaign.universe.Faction.Tag;
